Jess Carosel & image borders
Jess added the carosel onto the landing page (index.html), and commented out the code that was there before so it's not deleted. 
Jess also changed all of the .css files so that all images have a black border. This makes it look more uniform across the sites and the style goes with the font.